refactor(server): add exists by uuid check in delete game by uuid

+++ server/src/main/java/it/units/crossway/server/service/GameService.java
@@ -63,6 +63,8 @@ public class GameService {
 
     @Transactional
     public void deleteGame(String uuid) {
+        if (!gameRepository.existsByUuid(uuid))
+            throw new GameNotFoundException("Game with {uuid = " + uuid + "} not found");
         gameRepository.deleteByUuid(uuid);
     }
